botsy/src/platforms/instagram.py
import os
import logging
from dotenv import load_dotenv
from src.platforms.base_adapter import BasePlatformAdapter

# Load .env variables
load_dotenv()

class InstagramAdapter(BasePlatformAdapter):
    def __init__(self, bot):
        super().__init__(bot)
        self.access_token = None
        self.authenticate()

    def authenticate(self):
        # Attempt to load the Instagram access token from the environment
        self.access_token = os.getenv(f"{self.bot.name.upper()}_INSTAGRAM_ACCESS_TOKEN")
        if not self.access_token:
            # Fallback to a placeholder value if not provided
            self.access_token = "FAKE_INSTAGRAM_ACCESS_TOKEN"
        logging.info("InstagramAdapter: Authenticated using .env token (or placeholder).")

    def post(self, content: str):
        logging.info(f"InstagramAdapter: Posting content: {content}")
        return "instagram_post_id_12345"

    def comment(self, content: str, reply_to_id: str):
        logging.info(f"InstagramAdapter: Commenting on {reply_to_id}: {content}")
        return "instagram_comment_id_12345"

    def dm(self, recipient: str, message: str):
        logging.info(f"InstagramAdapter: Sending DM to {recipient}: {message}")
        return "instagram_dm_id_12345"